refactor(NameTest, NamesTest) : 도메인 패키지 변경
