+++ src/test/java/blackjack/domain/player/NameTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain.common;
+package blackjack.domain.player;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
