+++ src/test/java/blackjack/domain/player/NamesTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain.common;
+package blackjack.domain.player;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
